+++ SuperTerrain+/SuperRealism+/STPOrthographicCamera.cpp
@@ -7,21 +7,21 @@
 #include <glm/ext/matrix_clip_space.hpp>
 #include <glm/ext/matrix_transform.hpp>
 
-using glm::vec2;
-using glm::vec4;
-using glm::mat4;
+using glm::dvec2;
+using glm::dvec4;
+using glm::dmat4;
 
 using namespace SuperTerrainPlus::STPRealism;
 
 STPOrthographicCamera::STPOrthographicCamera
 	(const STPEnvironment::STPOrthographicCameraSetting& projection_props, const STPEnvironment::STPCameraSetting& camera_props) : 
-	STPCamera(camera_props), Frustum(projection_props), OrthographicProjection(glm::identity<mat4>()), ProjectionOutdated(true) {
+	STPCamera(camera_props), Frustum(projection_props), OrthographicProjection(glm::identity<dmat4>()), ProjectionOutdated(true) {
 	if (!this->Frustum.validate()) {
 		throw STPException::STPInvalidEnvironment("Orthographic projection setting not validated");
 	}
 }
 
-inline const mat4& STPOrthographicCamera::ortho() const {
+inline const dmat4& STPOrthographicCamera::ortho() const {
 	if (this->ProjectionOutdated) {
 		this->OrthographicProjection = glm::ortho(
 			this->Frustum.Left, 
@@ -38,11 +38,11 @@ inline const mat4& STPOrthographicCamera::ortho() const {
 	return this->OrthographicProjection;
 }
 
-const mat4& STPOrthographicCamera::projection() const {
+const dmat4& STPOrthographicCamera::projection() const {
 	return this->ortho();
 }
 
-mat4 STPOrthographicCamera::projection(float near, float far) const {
+dmat4 STPOrthographicCamera::projection(double near, double far) const {
 	return glm::ortho(
 		this->Frustum.Left,
 		this->Frustum.Right,
@@ -53,7 +53,7 @@ mat4 STPOrthographicCamera::projection(float near, float far) const {
 	);
 }
 
-void STPOrthographicCamera::reshape(vec4 side, vec2 depth) {
+void STPOrthographicCamera::reshape(dvec4 side, dvec2 depth) {
 	this->Frustum.Left = side.x;
 	this->Frustum.Right = side.y;
 	this->Frustum.Bottom = side.z;
